sync

+++ src/gView.GeoJsonService/DTOs/GetMapRequest.cs
@@ -11,7 +11,7 @@ public class GetMapRequest : BaseRequest
     public string Format { get; set; } = string.Empty;
     public bool Transparent { get; set; } = false;
     public double? Rotation { get; set; }
-    public int? Dpi { get; set; }
+    public float? Dpi { get; set; }
     public MapReponseFormat ResponseFormat { get; set; } = MapReponseFormat.Url;
 }
